+++ src/StorageBase/Action/PortalNode/Create/PortalNodeCreatePayload.php
@@ -8,41 +8,30 @@ use Heptacom\HeptaConnect\Dataset\Base\AttachmentCollection;
 use Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface;
 use Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait;
 use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalContract;
+use Heptacom\HeptaConnect\Portal\Base\Portal\PortalType;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Create\CreatePayloadInterface;
 
 final class PortalNodeCreatePayload implements CreatePayloadInterface, AttachmentAwareInterface
 {
     use AttachmentAwareTrait;
 
-    /**
-     * @var class-string<PortalContract>
-     */
-    protected string $portalClass;
+    protected PortalType $portalClass;
 
     protected ?string $alias;
 
-    /**
-     * @param class-string<PortalContract> $portalClass
-     */
-    public function __construct(string $portalClass, ?string $alias = null)
+    public function __construct(PortalType $portalClass, ?string $alias = null)
     {
         $this->attachments = new AttachmentCollection();
         $this->portalClass = $portalClass;
         $this->alias = $alias;
     }
 
-    /**
-     * @return class-string<PortalContract>
-     */
-    public function getPortalClass(): string
+    public function getPortalClass(): PortalType
     {
         return $this->portalClass;
     }
 
-    /**
-     * @param class-string<PortalContract> $portalClass
-     */
-    public function setPortalClass(string $portalClass): void
+    public function setPortalClass(PortalType $portalClass): void
     {
         $this->portalClass = $portalClass;
     }