+++ src/TestSuiteStorage/Action/JobTestContract.php
@@ -56,7 +56,7 @@ abstract class JobTestContract extends TestCase
         $jobStart = $facade->getJobStartAction();
 
         $firstPortalNode = $portalNodeCreate->create(new PortalNodeCreatePayloads([
-            new PortalNodeCreatePayload(PortalA::class),
+            new PortalNodeCreatePayload(PortalA::class()),
         ]))->first();
 
         static::assertInstanceOf(PortalNodeCreateResult::class, $firstPortalNode);